+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - Allow specifying transaction (span) ID via TransactionOptions/SpanOptions (#463)
  - module/apmzerolog: introduce zerolog log correlation and exception-tracking writer (#428)
  - module/apmelasticsearch: capture body for \_msearch, template and rollup search (#470)
+ - Ended Transactions/Spans may now be used as parents (#478)
 
 ## [v1.2.0](https://github.com/elastic/apm-agent-go/releases/tag/v1.2.0)